Finish migration of `_patching_infrastructure`
line_profiler/_child_process_profiling/_patching_infrastructure.py
    SingleModulePatch.submodule, .package
        Removed
    SingleModulePatch.module
        Now the first field, superseding `.submodule`
    Registry._default, .get_default()
        Removed

line_profiler/_child_process_profiling/multiprocessing_patches/
    __init__.py::get_registry()
        Rehabiilitated from `Registry.get_default()`
    __init__.py::_PATCHES
        Removed (`get_registry()` is cached)
    _pool_patch_helpers.py, *_patches.py
        Updated `SingleModulePatch` instantiations

tests/test_child_procs/_test_child_procs_utils.py, test_child_procs.py
    Replaced import of
    `line_profiler._child_process_profiling.multiprocessing_patches._PATCHES`
    with `.get_registry()`
